+++ modules/client/src/client-common/interfaces/common.ts
@@ -1,4 +1,5 @@
-export type SupportedNetworks = "mainnet" | "goerli";
+export const SupportedNetworksArray = ["mainnet", "goerli"] as const;
+export type SupportedNetworks = typeof SupportedNetworksArray[number];
 export type NetworkDeployment = {
   daoFactory: string;
   multisigRepo: string;